// src/generated/models/Cre2b_aplicacionsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cre2b_aplicacionsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cre2b_aplicacionsstatecode = keyof typeof Cre2b_aplicacionsstatecode;
export const Cre2b_aplicacionsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cre2b_aplicacionsstatuscode = keyof typeof Cre2b_aplicacionsstatuscode;

export interface Cre2b_aplicacionsBase {
  cre2b_aplicacionid: string;
  cre2b_link?: string;
  cre2b_nombre?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cre2b_aplicacionsstatecode;
  statuscode?: Cre2b_aplicacionsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cre2b_aplicacions extends Cre2b_aplicacionsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}