// src/generated/models/Cre2b_usuariosesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cre2b_usuariosesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cre2b_usuariosesstatecode = keyof typeof Cre2b_usuariosesstatecode;
export const Cre2b_usuariosesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cre2b_usuariosesstatuscode = keyof typeof Cre2b_usuariosesstatuscode;

export interface Cre2b_usuariosesBase {
  cre2b_departamento?: string;
  cre2b_email?: string;
  cre2b_nombre?: string;
  cre2b_posicion?: string;
  cre2b_usuariosid: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cre2b_usuariosesstatecode;
  statuscode?: Cre2b_usuariosesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cre2b_usuarioses extends Cre2b_usuariosesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}